requirements.txt
Django==3.2
django-allauth==0.63.6
django-cors-headers==4.4.0
django-crispy-forms==1.11.0
django-filter==2.4.0
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
phonenumbers==8.13.39
pillow==10.4.0
plotly==5.22.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
reportlab==4.2.2
seaborn==0.13.2
setuptools==65.5.0
six==1.16.0
sqlparse==0.5.0
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.1
